Add AboutModal component and related data
- Implement AboutModal for displaying project information and links.
- Create AboutData constants for project details and external links.
- Update SettingsMenu to include AboutModal functionality.
