+++ src/lib/constants/AboutData.ts
@@ -1,0 +1,15 @@
+export const aboutLinks = [
+  { label: 'WEBSITE', url: 'https://matt-bunch-dev.com' },
+  { label: 'GITHUB', url: 'https://github.com/MattBunch/Digital-Rain-Project' },
+  { label: 'LINKEDIN', url: 'https://www.linkedin.com/in/matt-bunch-dev/' },
+];
+
+export const projectInfo = {
+  version: '2.0.0',
+  description: 'Digital rain animation inspired by The Matrix, built with Svelte 5.',
+};
+
+export const licenseInfo = {
+  type: 'Apache 2.0',
+  url: 'https://www.apache.org/licenses/LICENSE-2.0',
+};
